Remove any usage of GenericType struct & type hack
This change finally gets rid of the `GenericType` struct as well as the
`unsafe` type casts it required. The new typing system for the channels
uses the [`Any trait`](https://doc.rust-lang.org/std/any/index.html) to
avoid any type hacks and allow for safe downcasting of the message
contents on the receiver side without introducing any newly generated
code (types can be inferred during downcast :tada:).

Also, this commit resolves a bug in the placement and usage of input
channels as they were not always being sorted 100% correctly and `pop`'d
from the vector in reverse order (lapidate me for that one).