Remove old backend optimizations, introduce hack for dead ends
I would declare this handling of `DeadArcs` as hack, since it would be nicer to
handle them natively during the code generation. The issue, however, is that
they pose a special case that needs to be detected at various points during
runtime. The old implementation already considered all arcs leading to
operator (0, 0) as DeadEnds so it was simpler to just remap the new operators to
the old scheme.