Include op id into fn signature to avoid redefinitions
The problem here is that when the `nth` operator, e.g. `nth_1_2` was used two or
more times throughout the code the function would be generated two times,
causing a compilation error.

The obvious fix for this would of course be to reuse the generated function. But
for now, this simple hotfix also adresses the problem.